// frontend/src/generated/types/menuDetailItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface MenuDetailItem {
  sysId?: string;
  linkSysId?: string;
  menuName?: string;
  menuPrice?: number;
  menuDescription?: string;
  optionUseYn?: string;
  linkSysId2?: string;
  useYn?: string;
  fileUlid?: string;
  ordNo?: number;
}