// frontend/src/generated/types/menuDetailRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { MenuDetailItem } from './menuDetailItem';

export interface MenuDetailRequest {
  newItems?: MenuDetailItem[];
  updateItems?: MenuDetailItem[];
  delItems?: MenuDetailItem[];
}